requirements.txt
aiohttp==3.9.1
google-generativeai==0.3.2
pathlib2==2.3.7
python-dotenv==1.0.1 